requirements.txt
aiofiles==23.2.1
aiogram==3.8.0
aiohttp==3.9.5
aioredis==2.0.1
aiosignal==1.3.1
annotated-types==0.7.0
arq==0.26.0
async-timeout==4.0.3
asyncio-mqtt==0.16.2
attrs==23.2.0
cachetools==5.3.3
certifi==2024.6.2
click==8.1.7
frozenlist==1.4.1
hiredis==2.3.2
idna==3.7
magic-filter==1.0.12
multidict==6.0.5
paho-mqtt==2.1.0
pydantic==2.7.4
pydantic_core==2.18.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==4.6.0
six==1.16.0
typing_extensions==4.12.2
yarl==1.9.4
